+++ app/api/__init__.py
@@ -1,0 +1,5 @@
+from .deps import get_db
+
+__all__ =[
+    "get_db"
+]
