+++ app/db/__init__.py
@@ -1,0 +1,7 @@
+from .base import Base
+from .database import SessionLocal, engine
+
+__all__ = ["Base", "SessionLocal", "engine"]
+
+# Import all the models to make sure they're registered with Base
+from app.models import job_models
