+++ app/db/base.py
@@ -1,0 +1,4 @@
+from sqlalchemy.orm import declarative_base
+
+# Base class for SQLAlchemy models
+Base = declarative_base()
