+++ app/models/job_models.py
@@ -3,8 +3,7 @@
 # defining how Python classes = database tables.
 
 from sqlalchemy import Column, Integer, String, Date
-from app.db.database import Base
-from app.schemas.job_schemas import ApplicationStatus
+from app.db import Base
 from app.schemas import ApplicationStatus
 from sqlalchemy import Enum as SQLEnum
